FEAT: Run OdinJob in Process instead of Pool (#12)
This change modified how OdinJob's are run, using a Multiprocessing Process instead of a Pool.

The Pool construct would hang if a process, started by the Pool, was killed by the kernel because of something like an OOM error. The Process construct returns gracefully in this situation, providing a non-zero exitcode to indicate what happened to the child process.

The Process construct does not offer simple return values like the Pool does. So this change also implements a process-safe shared Value object, to return run_delay_secs from the OdinJob.start function.