FIX: Odin Migration to reset ODS Process (#17)
During initial launch of ODIN Cubic ODS processing, the process ran into several errors that tainted the state of the Archive and Fact tables produced by the process.

This change slightly modified the archive and fact table processes, to avoid process failures and also executes a migration to reset the state of S3 objects for the process.