FIX: S&B API load_sids logic (#25)
This change fixes the logic used for the load_sids method of the ArchiveAFCAPI Job.

After deployment to PROD it was identified that incremental data loading was not occurring as expected. This change fixes the incremental loading logic of load_sids and also adds tests to ensure proper functioning of the method.