fix: use column names in dubdb views
This change updates the way duckdb file VIEWS are created. VIEWS were previously created with a `SELECT *` statement, this change selects the specific columns that are expected in the parquet files.

In the situation where columns have been added to parquet files, this change will allow the existing VIEWS to still function, instead of throwing an error because the `SELECT *` statement no longer matches the metadata stored in the duckdb file.

This change also updates the duckdb version being used and addes an environment tag to duckdb and data dictionary objects.