fix: TRANSACTION_HISTORY Archive OOM
This change is attempting to fix OOM errors being thowing by the ArchiveCubicQlikTable process for the TRANSACTION_HISTORY table.

This will stop generating CDC parquet files if a memory threshold is reached. This threshold should be reviewed in the future to confirm it had no negative effects on other parts of the pipeline.