fix: cubic REASON_DIMENSION migration
This changes adds a migration to resolve a type discrepency in the EDW.REASON_DIMENSION tables.

The recent cubic back-office upgrade introcued a type change for the REASON_DIMENSION table changing the `purse_type` column from an integer to a string. To handle this type change a new migration has been created to convert all parquet row groups to matching schemas for this table.

This change will resolve recent errors being thrown by odin because of miss-matched data types between parquet files in the same dataset.