fix: cubic REASON_DIMENSION migration (#48)
This changes adds a migration to resolve a type discrepancy in the EDW.REASON_DIMENSION tables.

The recent cubic back-office upgrade introduced a type change for the REASON_DIMENSION table changing the `purse_type` column from an integer to a string. To handle this type change a new migration has been created to convert all parquet row groups to matching schemas for this table.