fix: ODSFact TRANSACTION_HISTORY OOM (#51)
For the last few days, Odin has been throwing OOM errors when attempting to process the EDW.TRANSACTION_HISTORY table.

This error was traced back to the `ds_batched_join` function, which is designed to join a polars dataframe to a very large parquet dataset in a memory constrained environment. Unfortunately, the implementation was not as memory constrained as originally thought.

In this instance, the JOIN operations was matching to a very small number of records from almost every partition across the whole dataset. In converting a pyarrow RecordBatch to a polars Dataframe, for each JOIN match, the memory allocation from each matched RecordBatch was being leaked as the dataset was iterated through.

This conversion to Polars is only required to execute JOIN operations on key columns that may contain NULL values. If no NULL values need to be JOIN'ed on, the JOIN operation can occur entirely in pyarrow (using Tables) and the memory leak does not occur. The NON NULL operation is the primary JOIN operation that occurs, so the Polars conversion remains for the small subset of data needing this functionality.