feat: S&B API Restricted Process
This change adds a new ingestion process for the S&B AFC API.

This new process removes PPI from the public version of the `v_media` table and publishes a seperate "lookup" version of the table to a restrcited S3 path.

This change also includes a migration to delete the existing `v_media` exports and re-pull the data with PPI removed.