feat: S&B API Restricted Process (#52)
This change adds a new ingestion process for the S&B AFC API.

This new process removes PPI from the public version of the `v_media` table and publishes a separate "lookup" version of the table to a restricted S3 path.

This change also includes a migration to delete the existing `v_media` exports and re-pull the data with PPI removed.